frontend/requirements.txt
streamlit
requests
streamlit-option-menu
python-dotenv